model: consolidate client errors
Use a single type for errors returned by the ResourceProviderClient and
TestClient (since they are so similar anyway).
