+++ model/src/clients/mod.rs
@@ -1,8 +1,7 @@
+mod error;
 mod resource_provider_client;
 mod test_client;
 
-pub use resource_provider_client::{
-    Error as ResourceProviderClientError, ResourceProviderClient,
-    Result as ResourceProviderClientResult,
-};
-pub use test_client::{Error, Result, TestClient};
+pub use error::{Error, Result};
+pub use resource_provider_client::ResourceProviderClient;
+pub use test_client::TestClient;
