+++ model/src/clients/resource_provider_client.rs
@@ -1,7 +1,8 @@
+use super::error::{self, Result};
 use crate::constants::NAMESPACE;
 use crate::ResourceProvider;
 use kube::Api;
-use snafu::{ResultExt, Snafu};
+use snafu::ResultExt;
 
 /// An API Client for TestSys ResourceProvider CRD objects.
 ///
@@ -19,33 +20,13 @@ pub struct ResourceProviderClient {
     api: Api<ResourceProvider>,
 }
 
-/// The `Result` type returned by [`ResourceProviderClient`].
-pub type Result<T> = std::result::Result<T, Error>;
-
-/// The public error type for `ResourceProvider`.
-// TODO - consolidate error types https://github.com/bottlerocket-os/bottlerocket-test-system/issues/91
-#[derive(Debug, Snafu)]
-pub struct Error(InnerError);
-
-/// The private error type for `ResourceProvider`.
-#[derive(Debug, Snafu)]
-pub(crate) enum InnerError {
-    #[snafu(display("Error initializing the Kubernetes client: {}", source))]
-    Initialization { source: kube::Error },
-
-    #[snafu(display("Unable to {} {}: {}", method, what, source))]
-    KubeApiCall {
-        method: String,
-        what: String,
-        source: kube::Error,
-    },
-}
-
 impl ResourceProviderClient {
     /// Create a new [`ResourceProviderClient`] using either `KUBECONFIG` or the in-cluster
     /// environment variables.
     pub async fn new() -> Result<Self> {
-        let k8s_client = kube::Client::try_default().await.context(Initialization)?;
+        let k8s_client = kube::Client::try_default()
+            .await
+            .context(error::Initialization)?;
         Ok(Self {
             api: Api::<ResourceProvider>::namespaced(k8s_client, NAMESPACE),
         })
@@ -56,9 +37,13 @@ impl ResourceProviderClient {
     where
         S: AsRef<str>,
     {
-        Ok(self.api.get(name.as_ref()).await.context(KubeApiCall {
-            method: "get",
-            what: "resource provider",
-        })?)
+        Ok(self
+            .api
+            .get(name.as_ref())
+            .await
+            .context(error::KubeApiCall {
+                method: "get",
+                what: "resource provider",
+            })?)
     }
 }
